requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
click==8.2.1
discord.py==2.6.4
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.7.0
propcache==0.4.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
yarl==1.22.0
